Исправлен файл schema

+++ src/main/resources/schema.sql
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS friendship_type (
 
 -- Создание таблицы Consumer
 CREATE TABLE IF NOT EXISTS consumer (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
+    id INT PRIMARY KEY AUTO_INCREMENT,
     email VARCHAR(255) NOT NULL,
     login VARCHAR(100) NOT NULL,
     name VARCHAR(255),
@@ -42,7 +42,7 @@ CREATE TABLE IF NOT EXISTS consumer (
 
 -- Создание таблицы Film
 CREATE TABLE IF NOT EXISTS film (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT,
+    id INT PRIMARY KEY AUTO_INCREMENT,
     name VARCHAR(100) NOT NULL,
     description VARCHAR(200),
     release_date DATE NOT NULL,
@@ -53,8 +53,8 @@ CREATE TABLE IF NOT EXISTS film (
 
 -- Создание таблицы Film_Likes
 CREATE TABLE IF NOT EXISTS film_likes (
-    film_id BIGINT,
-    user_id BIGINT,
+    film_id INT,
+    user_id INT,
     PRIMARY KEY (film_id, user_id),
     FOREIGN KEY (film_id) REFERENCES film(id),
     FOREIGN KEY (user_id) REFERENCES consumer(id)
@@ -62,7 +62,7 @@ CREATE TABLE IF NOT EXISTS film_likes (
 
 -- Создание таблицы Film_Genre
 CREATE TABLE IF NOT EXISTS film_genre (
-    film_id BIGINT,
+    film_id INT,
     genre_id INT,
     PRIMARY KEY (film_id, genre_id),
     FOREIGN KEY (film_id) REFERENCES film(id),
@@ -71,8 +71,8 @@ CREATE TABLE IF NOT EXISTS film_genre (
 
 -- Создание таблицы User_Friends
 CREATE TABLE IF NOT EXISTS user_friends (
-    user_id BIGINT,
-    friend_id BIGINT,
+    user_id INT,
+    friend_id INT,
     friendship_type_id INT,
     PRIMARY KEY (user_id, friend_id),
     FOREIGN KEY (user_id) REFERENCES consumer(id),
